requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
camelot-py==1.0.9
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.17.0
nodeenv==1.10.0
numpy==2.0.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
protobuf==6.33.5
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pypdf==3.17.4
pypdfium2==5.5.0
pymupdf==1.26.5
openai==1.65.1
python-dotenv==1.0.1
pyright==1.1.408
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
redis==7.0.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.46
starlette==0.49.3
streamlit==1.50.0
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
pytest==8.4.2
httpx==0.28.1
psycopg2-binary==2.9.11
pgvector==0.4.2
